Extract DEM terrain grid sampler
